DTSBP-83: Returning number of copies requested (#828)
* DTSBP-83: Returning number of copies requested

* DTSBP-83: Added default copies values to 1

* DTSBP-83: Added default copies values to 1

+++ src/integrationTest/java/uk/gov/hmcts/reform/sendletter/services/LetterServiceTest.java
@@ -84,7 +84,7 @@ class LetterServiceTest {
         Letter result = letterRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Letter not found " + id.toString()));
 
-        assertEquals(14, result.getCopies());
+        assertEquals(14, result.getCopies().intValue());
         assertThat(result.isEncrypted()).isFalse();
         assertThat(result.getEncryptionKeyFingerprint()).isNull();
         PdfHelper.validateZippedPdf(result.getFileContent());
